refactor: move `merkle_sum_tree` module testing to `tests.rs`
